Move queries to separate module
Also fix some of their keys
